commit du 1er oct

+++ src/main/java/com/all4tic/suiviscolaire/dao/TypeSuiviDao.java
@@ -4,6 +4,6 @@ import com.all4tic.suiviscolaire.entities.TypeSuivi;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 @Repository
-public interface TypeSuiviDao extends CrudRepository<TypeSuivi, Long> {
+public interface TypeSuiviDao extends CrudRepository<TypeSuivi, Integer> {
 
 }
